Include the required linker script for the language data
Instead of relying on patching the Arduino's avr6.xn linker script
(which is required for the custom language data sections), provide the
linker script and use linker flags in boards.txt to use it.

This *greatly* simplifies the installation and usage of Arduino, and
also ensures the firmware is always built correctly, since a missing
section will result in a non-working firmware which is difficult to
diagnose.